requirements.txt
fastapi==0.115.0
python-socketio==5.11.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
